requirements.txt
streamlit>=1.24.0
pandas>=1.3.0
plotly>=5.3.0
numpy>=1.20.0
openpyxl>=3.1.0
python-dotenv>=0.19.0
google-generativeai>=0.3.0
